requirements.txt
alabaster==0.7.12
alembic==1.4.2
altair==4.1.0
argh==0.26.2
astor==0.8.1
astroid==2.4.2
astropy==4.0.1.post1
atomicwrites==1.4.0
attrs==19.3.0
autopep8==1.5.4
azure-core==1.8.0
azure-storage-blob==12.4.0
Babel==2.8.0
backcall==0.1.0
base58==2.0.1
bcrypt==3.2.0
bleach==3.1.4
blinker==1.4
blis==0.4.1
boto==2.49.0
boto3==1.14.48
botocore==1.17.48
brotlipy==0.7.0
cachetools==4.1.1
catalogue==1.0.0
catboost==0.24
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cloudpickle==1.5.0
colorama==0.4.3
colorlover==0.3.0
combo==0.1.1
confuse==1.3.0
cryptography==3.0
cufflinks==0.17.3
cycler==0.10.0
cymem==2.0.3
Cython==0.29.14
databricks-cli==0.11.0
datefinder==0.7.1
DateTime==4.3
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
docker==4.3.1
docutils==0.16
entrypoints==0.3
enum-compat==0.0.3
flake8==3.8.3
Flask==1.1.2
funcy==1.14
future==0.18.2
gensim==3.8.3
gitdb==4.0.5
GitPython==3.1.7
gorilla==0.3.0
graphviz==0.14.1
htmlmin==0.1.12
idna==2.10
ImageHash==4.1.0
imagesize==1.2.0
imbalanced-learn==0.7.0
importlib-metadata==1.6.0
iniconfig==1.0.1
intervaltree==3.0.2
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.15.1
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
keyring==21.2.1
kiwisolver==1.2.0
kmodes==0.10.2
lazy-object-proxy==1.4.3
lightgbm==2.3.1
llvmlite==0.34.0
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
missingno==0.4.2
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mlflow==1.10.0
mlxtend==0.17.3
more-itertools==8.4.0
msrest==0.6.18
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.6
networkx==2.5
nltk==3.5
notebook==6.0.3
numba==0.51.0
numexpr==2.7.1
numpy==1.18.1
numpydoc==1.1.0
oauthlib==3.1.0
packaging==20.4
pandas==1.0.3
pandas-profiling==2.8.0
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.7.0
pathtools==0.1.2
patsy==0.5.1
pexpect==4.8.0
phik==0.10.0
pickleshare==0.7.5
Pillow==7.2.0
plac==1.1.3
plotly==4.9.0
pluggy==0.13.1
preshed==3.0.2
prometheus-client==0.7.1
prometheus-flask-exporter==0.15.4
prompt-toolkit==3.0.5
protobuf==3.13.0
psutil==5.7.0
py==1.9.0
pyarrow==1.0.1
pycaret==2.0
pycodestyle==2.6.0
pycparser==2.20
pydeck==0.5.0b1
pydocstyle==5.1.0
pyflakes==2.2.0
Pygments==2.6.1
pyLDAvis==2.1.2
pylint==2.5.3
PyNaCl==1.4.0
pyod==0.8.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
pytest==6.0.1
python-dateutil==2.8.1
python-editor==1.0.4
python-jsonrpc-server==0.3.4
python-language-server==0.34.1
pytz==2020.1
PyWavelets==1.1.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==18.1.1
QDarkStyle==2.8.1
QtAwesome==0.7.2
qtconsole==4.7.5
QtPy==1.9.0
querystring-parser==1.2.4
regex==2020.6.8
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rope==0.17.0
Rtree==0.9.4
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
sip==4.19.13
six==1.14.0
smart-open==2.1.0
smmap==3.0.4
snowballstemmer==2.0.0
sortedcontainers==2.2.2
spacy==2.3.2
Sphinx==3.2.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
spyder==4.1.4
spyder-kernels==1.9.2
SQLAlchemy==1.3.13
sqlparse==0.3.1
srsly==1.0.2
statsmodels==0.11.1
streamlit==0.65.2
suod==0.0.4
tabulate==0.8.7
tangled-up-in-unicode==0.0.6
terminado==0.8.3
testpath==0.4.4
textblob==0.15.3
thinc==7.4.1
threadpoolctl==2.1.0
toml==0.10.1
toolz==0.10.0
tornado==6.0.4
tqdm==4.47.0
traitlets==4.3.3
typed-ast==1.4.1
tzlocal==2.1
ujson==1.35
umap-learn==0.4.6
urllib3==1.25.10
validators==0.18.0
visions==0.4.4
waitress==1.4.4
wasabi==0.7.1
watchdog==0.10.3
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
wordcloud==1.8.0
wrapt==1.11.2
xgboost==1.2.0
yapf==0.30.0
yellowbrick==1.1
zipp==3.1.0
zope.interface==5.1.0